requirements.txt
unstructured 
tiktoken
pinecone-client
pypdf
openai
langchain
langchain-community
langchain-openai
pandas
numpy
python-dotenv